Small ssh and ip changes
Add banner_timeout to paramiko connect in order to wait for inactive
host. Otherwise paramiko raises double exception which cannot be
handled.
Also replace None to localhost ip in dut.ip in Local mode, as it's real
Dut ip in Local mode and can be used for network communication in some
cases.

Signed-off-by: Oleksandr Shchirskyi <oleksand.shchirskyi@intel.com>